src/client/requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.19.0 